+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 
 ## mini.completion
 
+- BREAKING FEATURE: add snippet support. By default uses 'mini.snippets' to manage snippet session (if enabled, **highly recommended), falls back to `vim.snippet` on Neovim>=0.10. See "Snippets" section in `:h MiniCompletion` for more details.<br>
+  This is a breaking change mostly because items with `Snippet` kind are no longer filtered out by default.
 - FEATURE: add scrolling in info and signature window. By default can be done with `<C-f>` / `<C-b>` when target window is shown. Can be configured via `mappings.scroll_down` and `mappings.scroll_up` config options.
 - FEATURE: respect `isIncomplete` in LSP completion response and immediately force new completion request on the next key press.
 - FEATURE: both info and signature help windows now use tree-sitter highlighting: